network_interface.py updates
Created module documentation
Update logger to use local logger
created set_devices_queue as interface for others modules to call this module sending a list of devices. function transfor the list to a queue and send it to producer_manager to execute commands on devices. Returns a list of the devices